+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 # [Unreleased]
 # Added
 - Added UMM option to change spells known per level.
+- Added UMM option to disable spellbooks for each Isekai archetype.
 - Added UMM option to change Isekai Protagonist to have the Shifter appearance.
 
 # [5.2.3] - 2023-05-30